+++ stories/pages/dropdown/default.stories.js
@@ -4,11 +4,18 @@ import { template } from '../../templates/dropdown/default.template';
 import documentation from '../../documentation/dropdown.md'
 import { PopperPlacement, TriggerType } from '../../../src/components/popper/types';
 
+let notes = documentation;
+notes = notes.replace('{{DEFAULT_PLACEMENT}}', PopperPlacement.bottomStart);
+notes = notes.replace('{{AVAILABLE_PLACEMENTS}}', Object.values(PopperPlacement).join(', '));
+
+notes = notes.replace('{{DEFAULT_TRIGGER}}', TriggerType.click);
+notes = notes.replace('{{AVAILABLE_TRIGGERS}}', Object.values(TriggerType).join(', '));
+
 storiesOf('Components|Dropdown', module).add('Default', () => ({
   template,
   props: {
     placement: {
-      default: select('placement', Object.values(PopperPlacement), PopperPlacement.bottom)
+      default: select('placement', Object.values(PopperPlacement), PopperPlacement.bottomStart)
     },
     trigger: {
       default: select('trigger', Object.values(TriggerType), TriggerType.click)
@@ -51,5 +58,5 @@ storiesOf('Components|Dropdown', module).add('Default', () => ({
     return {}
   },
 }), {
-  notes: { markdown: documentation },
+  notes: { markdown: notes },
 });