+++ stories/pages/popper/default.stories.js
@@ -6,10 +6,10 @@ import { PopperPlacement, TriggerType } from "../../../src/components/popper/typ
 
 let notes = documentation;
 notes = notes.replace('{{DEFAULT_PLACEMENT}}', PopperPlacement.auto);
-notes = notes.replace('{{AVAILABLE_PLACEMENTS}}', Object.values(PopperPlacement).join(' , '));
+notes = notes.replace('{{AVAILABLE_PLACEMENTS}}', Object.values(PopperPlacement).join(', '));
 
 notes = notes.replace('{{DEFAULT_TRIGGER}}', TriggerType.hover);
-notes = notes.replace('{{AVAILABLE_TRIGGERS}}', Object.values(TriggerType).join(' , '));
+notes = notes.replace('{{AVAILABLE_TRIGGERS}}', Object.values(TriggerType).join(', '));
 
 storiesOf('Components|Popper', module).add('Default', () => ({
   template,