+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Now you can set a suffix-icon near the default arrow icon
 * Added validation for `Input`
 * Added `optionClass` prop to `Select`
 * Added `size` prop to `Autocomplete` and `Select`
+* `Dropdown` placement was changed from `bottom` to `bottom-start`. Same in `Autocomplete` and `Select` components
 
 ### Maintenance
 * Updated `Notification` component. Do not show browser notifications when window is active
@@ -22,6 +23,7 @@ Now you can set a suffix-icon near the default arrow icon
 * Fixed server sorting in `Table` component
 * `Scrollbar`'s caret containers was updated with z-index
 * Fixed `Input`'s icon sizes with different value of `size` prop
+* Fixed `Dropdown` documentation page